# app/Models/Timesheet.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Timesheet extends Model
{
    use HasFactory;

    protected $fillable = [
        'user_id',
        'type',
        'day_in',
        'day_out',
    ];

    protected $casts = [
        'day_in' => 'datetime',
        'day_out' => 'datetime',
    ];

    public function user()
    {
        return $this->belongsTo(User::class);
    }
}